refactor project with MVVM structure

+++ app/src/main/java/com/example/Bank/models/carts/DataCarts.kt
@@ -7,9 +7,9 @@ data class DataCarts(
     var cart_items: List<ListOfCartItems>,
 
     @SerializedName("sub_total")
-    var sub_total: Int,
+    var sub_total: Double,
 
     @SerializedName("total")
-    var total: Int,
+    var total: Double,
 
 )
